counter2_0.py and unit_tests_counter2_0.py have been renamed to counter.py and unit_tests_counter.py
